ADD: seeds and erd change

+++ db/migrate/20221007040158_create_albums.rb
@@ -2,7 +2,7 @@ class CreateAlbums < ActiveRecord::Migration[7.0]
   def change
     create_table :albums do |t|
       t.string :name, null: false
-      t.decimal :price, null: false, precision: 10, scale: 2
+      t.integer :price, null: false, precision: 10, scale: 2
       t.references :artist, null: false, foreign_key: true
 
       t.timestamps
